logger manager (#45909)
uniform logger manager in FleetAPI.
hidde API under distributed/utils which users don't need.
