+++ python/paddle/distributed/fleet/utils/hybrid_parallel_util.py
@@ -14,7 +14,6 @@
 import os
 import six
 import numpy as np
-import warnings
 
 from paddle import framework
 import paddle
